fixing type int errors

+++ ft_itoa.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nabdelba <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nabdelba <nabdelba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/16 14:17:14 by nabdelba          #+#    #+#             */
-/*   Updated: 2019/04/17 21:45:34 by nabdelba         ###   ########.fr       */
+/*   Updated: 2019/08/28 15:47:01 by nabdelba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char		*ft_itoa(int n)
+char		*ft_itoa(intmax_t n)
 {
-	char	*str;
-	int		i;
-	int		num;
+	char			*str;
+	int				i;
+	uintmax_t		num;
 
 	i = 0;
 	num = n;
